supplicant: rework wpa_supplicant handling
Avoid GDBusProxy, instead use GDBusConnection directly. I very much
prefer this because that way we have explicit control over what happens
on D-Bus. With GDBusProxy this is hidden under another layer of complex
code. The hardest part when using a D-Bus interface is to manage the
state. GDBusProxy contain state about the D-Bus interface and duplicate
the state that we track. This makes it hard to reason about things.

Rework creation of NMSupplicantInterface. Previously, a
NMSupplicantInterface had multiple initialization states.
In particular, the first state would not yet tie the interface
to a certain D-Bus object path. Instead, NMSupplicantInterface
would try and retry to create the D-Bus object.
Now, NMSupplicantManager has an asynchronous method to create interface
instances. The manager only creates a interface instance after the D-Bus
path is known. That means, now one NMSupplicantInterface instance is now
strongly tied to a name-owner and D-Bus path. These cannot change.

It follows that the state of NMSupplicantInterface can only go from STARTING
via supplicant states to DOWN. And never back. That was already
previously the case that the state from DOWN was final and once the 3 initial
states were passed, it would never go back to the initial state. What
change is that this is now more strict. The 3 initialization states are combined
and the progression is very clear.

I think the tighter state handling simplifies users of NMSupplicantInterface.
See for example "nm-device-ethernet.c". It's still complicated, because handling
state is fundamentally difficult.